+++ src/frontend/komponenter/Behandling/Inngangsvilkår/Samliv/typer.ts
@@ -48,3 +48,10 @@ export const SagtOppEllerRedusertTilTekst: Record<ESagtOppEllerRedusert, string>
     redusertStilling: 'Ja, jeg har redusert arbeidstiden',
     nei: 'Nei',
 };
+
+export interface ISivilstandsplaner {
+    harPlaner?: boolean;
+    fraDato?: string;
+    vordendeSamboerEktefelle?: IPersonDetaljer;
+}
+
