--- src/frontend/komponenter/Behandling/Inngangsvilkår/config/DelvilkårConfig.ts
@@ -1,15 +1,0 @@
-import { DelvilkårType } from '../vilkår';
-
-interface IDelvilkårConfig {
-    harBegrunnelse?: boolean;
-}
-
-type DelvilkårConfig = {
-    [key in DelvilkårType]?: IDelvilkårConfig;
-};
-
-export const DelvilkårConfig: DelvilkårConfig = {
-    DOKUMENTERT_EKTESKAP: {},
-    DOKUMENTERT_SEPARASJON_ELLER_SKILSMISSE: {},
-    KRAV_SIVILSTAND: {},
-};
