+++ src/frontend/komponenter/Behandling/Vurdering/hjelpetekstconfig.tsx
@@ -1,0 +1,6 @@
+import NæreBoforholdHjelpetekst from "../Inngangsvilkår/Aleneomsorg/NæreBoforholdHjelpetekst";
+import React from "react";
+
+export const hjelpeTekstConfig: Record<string, React.FC> = {
+    NÆRE_BOFORHOLD : NæreBoforholdHjelpetekst
+}
